Добавлен пакет control

+++ src/main/java/kz/mix/e804/localization/resourcebundles/ResBundle_it_IT.java
@@ -3,6 +3,7 @@ package kz.mix.e804.localization.resourcebundles;
 
 import java.util.ListResourceBundle;
 
+// Italian version
 public class ResBundle_it_IT extends ListResourceBundle {
     public Object[][] getContents() {
         return contents;
@@ -10,7 +11,7 @@ public class ResBundle_it_IT extends ListResourceBundle {
 
     static final Object[][] contents = {
             { "MovieName", "Che Bella Giornata" },
-            { "GrossRevenue", (Long) 43000000L }, // in euros
-            { "Year", (Integer)2011 }
+            { "GrossRevenue", 43000000L}, // in euros
+            { "Year", 2011}
     };
 }
